requirements.txt
asgiref==3.3.1
Django==3.1.5
pytz==2020.5
sqlparse==0.4.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
djangorestframework==3.12.2
django-cors-headers==3.7.0